v.1.1.0, code & structure refactoring, remove old comments from github, fix SQ issues
