+++ src/main/java/org/perpectiveteam/plugins/aisummarize/AiSummarizePlugin.java
@@ -4,17 +4,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.perpectiveteam.plugins.aisummarize.ai.AIPromptBuilder;
-import org.perpectiveteam.plugins.aisummarize.config.AISummarizeConfigProperties;
-import org.perpectiveteam.plugins.aisummarize.config.AiSummarizeConfig;
-import org.perpectiveteam.plugins.aisummarize.config.AiSummarizeProperySensor;
-import org.perpectiveteam.plugins.aisummarize.hooks.PostProjectAnalysisSummarize;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.PostAnalysisIssueVisitor;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.bitbucket.cloud.BitbucketCloudClientFactory;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.bitbucket.BitbucketConfiguration;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.bitbucket.HttpClientBuilderFactory;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.github.GitHubClientFactory;
-import org.perpectiveteam.plugins.aisummarize.summarize.SummarizeExecutorFactory;
-import org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.ALMClientFactory;
+import org.perpectiveteam.plugins.aisummarize.config.SummarizeConfigProperties;
+import org.perpectiveteam.plugins.aisummarize.config.SummarizeConfig;
+import org.perpectiveteam.plugins.aisummarize.config.SummarizeProperySensor;
+import org.perpectiveteam.plugins.aisummarize.ce.PostProjectAnalysisSummarize;
+import org.perpectiveteam.plugins.aisummarize.analysis.PostAnalysisIssueVisitor;
+import org.perpectiveteam.plugins.aisummarize.almclient.bitbucket.cloud.BitbucketCloudClientFactory;
+import org.perpectiveteam.plugins.aisummarize.almclient.bitbucket.BitbucketConfiguration;
+import org.perpectiveteam.plugins.aisummarize.almclient.HttpClientBuilderFactory;
+import org.perpectiveteam.plugins.aisummarize.almclient.github.GitHubClientFactory;
+import org.perpectiveteam.plugins.aisummarize.ce.SummarizeExecutorFactory;
+import org.perpectiveteam.plugins.aisummarize.almclient.ALMClientFactory;
 import org.sonar.api.Plugin;
 import org.sonar.api.SonarQubeSide;
 import org.sonar.core.extension.CoreExtension;
@@ -37,22 +37,22 @@ public class AiSummarizePlugin implements Plugin, CoreExtension {
                     HttpClientBuilderFactory.class,
                     BitbucketCloudClientFactory.class,
                     SummarizeExecutorFactory.class,
-                    AiSummarizeConfig.class,
+                    SummarizeConfig.class,
                     AIPromptBuilder.class
                 )
             );
-            extensions.addAll(AISummarizeConfigProperties.all());
+            extensions.addAll(SummarizeConfigProperties.all());
             context.addExtensions(extensions);
         } else if (SonarQubeSide.SERVER == context.getRuntime().getSonarQubeSide()) {
-            context.addExtensions(AISummarizeConfigProperties.all());
+            context.addExtensions(SummarizeConfigProperties.all());
         }
     }
 
     @Override
     public void define(Plugin.Context context) {
         context.addExtensions(
-                AiSummarizeConfig.class,
-                AiSummarizeProperySensor.class
+                SummarizeConfig.class,
+                SummarizeProperySensor.class
         );
     }
 }
