+++ src/main/java/org/perpectiveteam/plugins/aisummarize/ai/AIPromptBuilder.java
@@ -4,44 +4,44 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.perpectiveteam.plugins.aisummarize.config.AiSummarizeConfig;
+import org.perpectiveteam.plugins.aisummarize.config.SummarizeConfig;
 import org.perpectiveteam.plugins.aisummarize.pullrequest.prdto.FileDiff;
 import org.sonar.api.ce.ComputeEngineSide;
 import org.sonar.api.server.ServerSide;
-import org.sonar.api.utils.log.Logger;
-import org.sonar.api.utils.log.Loggers;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @ServerSide
 @ComputeEngineSide
 public class AIPromptBuilder {
-    private final AiSummarizeConfig aiSummarizeConfig;
+    private final SummarizeConfig aiSummarizeConfig;
+    private static final String DELIMITER = "========================================\n";
 
-    public AIPromptBuilder(AiSummarizeConfig aiSummarizeConfig) {
+    public AIPromptBuilder(SummarizeConfig aiSummarizeConfig) {
         this.aiSummarizeConfig = aiSummarizeConfig;
     }
 
-    private static final Logger LOG = Loggers.get(AIPromptBuilder.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(AIPromptBuilder.class);
 
     public String buildPrompt(List<FileDiff> fileDiffs) {
-        //TODO: include SQ issues
         ParsedTemplate parsed = parseTemplate(aiSummarizeConfig);
-        LOG.info("Building AI prompt for {} files", fileDiffs.size());
+        LOGGER.info("Building AI prompt for {} files", fileDiffs.size());
         StringBuilder sb = new StringBuilder();
         sb.append(parsed.intro);
-        sb.append("========================================\n");
+        sb.append(DELIMITER);
 
         int validFileCount = 0;
         for (FileDiff fileDiff : fileDiffs) {
             if (fileDiff.rawContent == null || fileDiff.filePath.isEmpty() || fileDiff.rawContent.isEmpty()) {
                 fileDiff.rawContent = "There is no previous version, probably a new file";
-                LOG.debug("Missing previous file version for: " + fileDiff.filePath);
+                LOGGER.debug("Missing previous file version for: {}", fileDiff.filePath);
             }
             if (fileDiff.filePath.isEmpty() || fileDiff.changes.isEmpty()) {
-                LOG.debug("Skipping file with missing data: {}", fileDiff.filePath);
+                LOGGER.debug("Skipping file with missing data: {}", fileDiff.filePath);
                 continue;
             }
             validFileCount++;
-            sb.append("========================================\n");
+            sb.append(DELIMITER);
             sb.append("Filename: ").append(fileDiff.filePath).append("\n\n");
 
             sb.append("Original Content:\n");
@@ -55,8 +55,8 @@ public class AIPromptBuilder {
             sb.append("-----\n\n");
         }
 
-        LOG.info("Included {} valid files in the prompt", validFileCount);
-        sb.append("========================================\n");
+        LOGGER.info("Included {} valid files in the prompt", validFileCount);
+        sb.append(DELIMITER);
         sb.append(parsed.after);
         return sb.toString();
     }
@@ -71,11 +71,11 @@ public class AIPromptBuilder {
         }
     }
 
-    public static ParsedTemplate parseTemplate(AiSummarizeConfig aiSummarizeConfig) {
+    public static ParsedTemplate parseTemplate(SummarizeConfig aiSummarizeConfig) {
         String template = aiSummarizeConfig.getAiPromptTemplate();
 
-        Pattern introPattern = Pattern.compile(AiSummarizeConfig.INTRO_REGEXP, Pattern.DOTALL);
-        Pattern afterPattern = Pattern.compile(AiSummarizeConfig.AFTER_REGEXP, Pattern.DOTALL);
+        Pattern introPattern = Pattern.compile(SummarizeConfig.INTRO_REGEXP, Pattern.DOTALL);
+        Pattern afterPattern = Pattern.compile(SummarizeConfig.AFTER_REGEXP, Pattern.DOTALL);
 
         Matcher introMatcher = introPattern.matcher(template);
         Matcher afterMatcher = afterPattern.matcher(template);
