+++ src/main/java/org/perpectiveteam/plugins/aisummarize/almclient/HttpClientBuilderFactory.java
@@ -1,0 +1,13 @@
+package org.perpectiveteam.plugins.aisummarize.almclient;
+
+import okhttp3.OkHttpClient;
+import org.sonar.api.ce.ComputeEngineSide;
+import org.sonar.api.server.ServerSide;
+
+@ServerSide
+@ComputeEngineSide
+public class HttpClientBuilderFactory {
+    public OkHttpClient.Builder createClientBuilder() {
+        return new OkHttpClient.Builder();
+    }
+}
