+++ src/main/java/org/perpectiveteam/plugins/aisummarize/almclient/bitbucket/BitbucketCloudException.java
@@ -1,0 +1,7 @@
+package org.perpectiveteam.plugins.aisummarize.almclient.bitbucket;
+
+public class BitbucketCloudException extends RuntimeException {
+    public BitbucketCloudException(String message) {
+        super(message);
+    }
+}
