+++ src/main/java/org/perpectiveteam/plugins/aisummarize/almclient/github/GithubClientException.java
@@ -1,0 +1,7 @@
+package org.perpectiveteam.plugins.aisummarize.almclient.github;
+
+public class GithubClientException extends RuntimeException {
+  public GithubClientException(String message) {
+    super(message);
+  }
+}
