+++ src/main/java/org/perpectiveteam/plugins/aisummarize/ce/SummarizeExecutorException.java
@@ -1,0 +1,7 @@
+package org.perpectiveteam.plugins.aisummarize.ce;
+
+public class SummarizeExecutorException extends RuntimeException {
+    public SummarizeExecutorException(String message) {
+        super(message);
+    }
+}
