+++ src/main/java/org/perpectiveteam/plugins/aisummarize/config/options/AiClientsList.java
@@ -7,6 +7,10 @@ public class AiClientsList {
     public static final String OPENAI = "openai";
     public static final String OPENROUTER = "openrouter";
 
+    private AiClientsList() {
+        throw new IllegalStateException("Utility class");
+    }
+
     public static List<String> options() {
         return Arrays.asList(OPENAI, OPENROUTER);
     }
