--- src/main/java/org/perpectiveteam/plugins/aisummarize/pullrequest/almclient/bitbucket/BitbucketConfiguration.java
@@ -1,20 +1,0 @@
-package org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.bitbucket;
-
-public class BitbucketConfiguration {
-
-    private final String repository;
-    private final String project;
-
-    public BitbucketConfiguration(String project, String repository) {
-        this.repository = repository;
-        this.project = project;
-    }
-
-    public String getRepository() {
-        return repository;
-    }
-
-    public String getProject() {
-        return project;
-    }
-}
