--- src/main/java/org/perpectiveteam/plugins/aisummarize/pullrequest/almclient/bitbucket/cloud/comment/BitbucketSummarizeComment.java
@@ -1,19 +1,0 @@
-package org.perpectiveteam.plugins.aisummarize.pullrequest.almclient.bitbucket.cloud.comment;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class BitbucketSummarizeComment {
-    @JsonProperty("content")
-    private final BitbucketCommentContent content;
-
-
-    public BitbucketSummarizeComment(BitbucketCommentContent content) {
-        this.content = content;
-    }
-
-
-    public BitbucketCommentContent getContent() {
-        return content;
-    }
-
-}
